Implement UVC/Linux compatible way to query BNO information
Some UVC controls are allowed to AUTO_UPDATE, while others are not.
Linux enforces this, while Windows does not - but there is nothing
stopping Windows to change that behavior, as it is the right thing to do
according to the specification.

Therefore this patch changes the firmware to use an AUTO_UPDATE control,
PANTILT, on Linux to query BNO values. This has the added benefit of
only needing 1-2 queries to get all BNO quaternions, as opposed to 4.
The previous code remains in place to not break older software on
Windows, but we also implement a way to query the current device
firmware ABI version, so the userspace software knows which properties
to use and can switch to using new properties transparently in future.